+++ packages/swc-plugin/src/shared/utils/js/stylex/tests/stylex_create_theme_test.rs
@@ -28,7 +28,7 @@ mod stylex_create_theme {
   }
 
   fn exprected_result_factory(
-    injected_styles: &[(&str, (&str, f32))],
+    injected_styles: &[(&str, (&str, f64))],
   ) -> IndexMap<String, Box<InjectableStyle>> {
     let mut expected_injected_styles = IndexMap::new();
