+++ packages/swc-plugin/src/shared/utils/js/stylex/tests/stylex_keyframes_test.rs
@@ -36,7 +36,7 @@ mod stylex_keyframes {
   }
 
   fn exprected_css_result_factory(
-    injected_styles: &[(&str, (&str, f32))],
+    injected_styles: &[(&str, (&str, f64))],
   ) -> IndexMap<String, InjectableStyle> {
     let mut expected_injected_styles = IndexMap::new();
